// src/app/modules/shared/models/offer.ts
export class Offer {
    id: number | undefined;
    image: string | undefined;
    title: string | undefined;
    description: string | undefined;
    type: number | undefined;
    adType: number | undefined;
    rooms: number | undefined;
    baths: number | undefined;
    price: number | undefined;
}